backend/requirements.txt
Flask==3.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
docutils==0.21.2
idna==3.10
Kivy==2.3.0
Pygments==2.18.0
requests==2.32.3
urllib3==2.2.3
psycopg2-binary==2.9.10